wallet: api server only behind the feature
