+++ xelis_wallet/src/api/server/mod.rs
@@ -1,0 +1,57 @@
+mod rpc_server;
+mod xswd_server;
+
+pub use rpc_server::{
+    WalletRpcServer,
+    WalletRpcServerShared,
+    AuthConfig
+};
+
+use serde::Serialize;
+use xelis_common::{
+    api::wallet::NotifyEvent,
+    rpc_server::WebSocketServerHandler
+};
+pub use xswd_server::{
+    XSWDServer,
+    XSWDWebSocketHandler
+};
+
+use crate::api::XSWDHandler;
+
+
+pub enum APIServer<W>
+where
+    W: Clone + Send + Sync + XSWDHandler + 'static
+{
+    RPCServer(WalletRpcServerShared<W>),
+    XSWD(XSWDServer<W>)
+}
+
+impl<W> APIServer<W>
+where
+    W: Clone + Send + Sync + XSWDHandler + 'static
+{
+    pub async fn notify_event<V: Serialize>(&self, event: &NotifyEvent, value: &V) {
+        let json = serde_json::to_value(value).unwrap();
+        match self {
+            APIServer::RPCServer(server) => {
+                server.get_websocket().get_handler().notify(event, json).await;
+            },
+            APIServer::XSWD(xswd) => {
+                xswd.get_handler().notify(event, json).await;
+            }
+        }
+    }
+
+    pub async fn stop(self) {
+        match self {
+            APIServer::RPCServer(server) => {
+                server.stop().await;
+            },
+            APIServer::XSWD(xswd) => {
+                xswd.stop().await;
+            }
+        }
+    }
+}
