+++ xelis_wallet/src/api/server/rpc_server.rs
